Удаление ненужных файлов и функций
