+++ js/common/params.js
@@ -1,18 +1,4 @@
-// Количество постов/фото
-const POSTS_QUANTITY = 25;
-
-// Интервал кол-ва лайков
-const LIKES_RANGE = {
-  min: 15,
-  max: 200
-};
-
-// Интервал кол-ва комментариев
-const COMMENTS_RANGE = {
-  min: 0,
-  max: 200
-};
-
+// Длина комментария
 const COMMENT_LENGTHS = {
   min: 20,
   max: 140
@@ -75,4 +61,4 @@ const IMAGE_EFFECTS = {
   }
 };
 
-export {POSTS_QUANTITY, LIKES_RANGE, COMMENTS_RANGE, COMMENT_LENGTHS, ZOOM_RANGE, IMAGE_EFFECTS};
+export {COMMENT_LENGTHS, ZOOM_RANGE, IMAGE_EFFECTS};
