+++ js/posts/thumbnailsRender.js
@@ -1,8 +1,10 @@
-import {getPosts, showError} from './api/api.js';
+import {getPosts} from '../api/api.js';
+import {showError} from '../common/messages.js';
 
 const pictureTemplate = document.querySelector('#picture').content.querySelector('.picture');
 const pictures = document.querySelector('.pictures');
 
+// Отрисовка постов других пользователей
 getPosts((posts) => {
   posts.forEach((post) => {
     const thumbnail = pictureTemplate.cloneNode(true);
